+++ src/main/java/domain/deck/RandomDeckGenerator.java
@@ -1,12 +1,18 @@
 package domain.deck;
 
-import domain.Card;
-import domain.Denomination;
-import domain.Suits;
+import domain.card.Card;
+import domain.card.Denomination;
+import domain.card.Suits;
 import java.util.Collections;
 import java.util.Stack;
 
 public class RandomDeckGenerator implements DeckGenerator {
+    private static void addCard(Stack<Card> cards, Suits suits) {
+        for (Denomination denomination : Denomination.values()) {
+            cards.push(new Card(denomination, suits));
+        }
+    }
+
     @Override
     public Deck generateDeck() {
         Stack<Card> cards = new Stack<>();
@@ -15,11 +21,5 @@ public class RandomDeckGenerator implements DeckGenerator {
         }
         Collections.shuffle(cards);
         return new Deck(cards);
-    }
-
-    private static void addCard(Stack<Card> cards, Suits suits) {
-        for (Denomination denomination : Denomination.values()) {
-            cards.push(new Card(denomination, suits));
-        }
     }
 }
