+++ src/test/java/domain/card/CardsTest.java
@@ -1,11 +1,11 @@
-package domain;
+package domain.card;
 
-import static domain.Denomination.ACE;
-import static domain.Denomination.FIVE;
-import static domain.Denomination.FOUR;
-import static domain.Denomination.SIX;
-import static domain.Denomination.THREE;
-import static domain.Denomination.TWO;
+import static domain.card.Denomination.ACE;
+import static domain.card.Denomination.FIVE;
+import static domain.card.Denomination.FOUR;
+import static domain.card.Denomination.SIX;
+import static domain.card.Denomination.THREE;
+import static domain.card.Denomination.TWO;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
