+++ src/test/java/domain/deck/DeckTest.java
@@ -1,9 +1,10 @@
-package domain;
+package domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.deck.Deck;
-import domain.deck.DefaultDeckGenerator;
+import domain.card.Card;
+import domain.card.Denomination;
+import domain.card.Suits;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
