+++ src/test/java/domain/game/BlackJackTest.java
@@ -1,16 +1,19 @@
-package domain;
+package domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.Denomination;
+import domain.card.Suits;
 import domain.deck.DefaultDeckGenerator;
 import domain.user.Dealer;
 import domain.user.Player;
+import domain.user.Users;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import result.Result;
 
 public class BlackJackTest {
     private Users users;
