+++ src/test/java/domain/user/PlayerTest.java
@@ -1,10 +1,12 @@
-package domain;
+package domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.user.Player;
+import domain.card.Card;
+import domain.card.Denomination;
+import domain.card.Suits;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
