+++ src/test/java/domain/user/UserTest.java
@@ -1,9 +1,10 @@
-package domain;
+package domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.user.Dealer;
-import domain.user.User;
+import domain.card.Card;
+import domain.card.Denomination;
+import domain.card.Suits;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
